+++ proguardprocessor/internal/metadatatest/generated_telemetrytest_test.go
@@ -10,6 +10,7 @@ import (
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
 
+	"github.com/honeycombio/opentelemetry-collector-symbolicator/proguardprocessor/internal/metadata"
 	"go.opentelemetry.io/collector/component/componenttest"
 )